[refactor]: 테스트 패키지 이름 수정
- application -> presentation으로 이름 수정
- repository -> persistence로 이름 수정

관련 이슈: #7
